refactor(mcp): split examples into individual files under examples/mcp/
One file per example:
- tool_discovery.py — basic tool discovery
- tool_call_with_retry.py — retry and timeout options
- per_tool_overrides.py — with_options() on individual tools
- interactive_assistant.py — pause/resume with MCP tools
- agent_with_mcp_tools.py — agent() + mcp() integration
- authenticated_server.py — bearer token via secret store
